chore: cleanup test mocks
