+++ packages/fiber/__mocks__/expo-file-system.ts
@@ -1,0 +1,1 @@
+export const readAsStringAsync = async () => new Promise((res) => res(''))
