+++ packages/fiber/tests/setupTests.ts
@@ -1,8 +1,6 @@
 import { createWebGLContext } from '@react-three/test-renderer/src/createWebGLContext'
 import * as THREE from 'three'
 import { extend } from '../src'
-import * as React from 'react'
-import { ViewProps, LayoutChangeEvent } from 'react-native'
 
 // Mock scheduler to test React features
 jest.mock('scheduler', () => require('scheduler/unstable_mock'))
@@ -15,55 +13,3 @@ HTMLCanvasElement.prototype.getContext = function () {
 
 // Extend catalogue for render API in tests
 extend(THREE)
-
-// Mock native dependencies for native
-jest.mock('react-native', () => ({
-  StyleSheet: {},
-  View: React.memo(
-    React.forwardRef(({ onLayout, ...props }: ViewProps, ref) => {
-      React.useLayoutEffect(() => {
-        onLayout?.({
-          nativeEvent: {
-            layout: {
-              x: 0,
-              y: 0,
-              width: 1280,
-              height: 800,
-            },
-          },
-        } as LayoutChangeEvent)
-
-        ref = { current: { props } }
-      }, [])
-
-      return null
-    }),
-  ),
-}))
-jest.mock('react-native/Libraries/Pressability/Pressability.js', () => ({}))
-class Asset {
-  name = 'test asset'
-  type = 'glb'
-  hash = null
-  uri = 'test://null'
-  localUri = 'test://null'
-  width = 800
-  height = 400
-  static fromURI = () => this
-  static fromModule = () => this
-  static downloadAsync = async () => new Promise((res) => res(this))
-}
-jest.mock('expo-asset', () => ({ Asset }))
-jest.mock('expo-file-system', () => ({
-  readAsStringAsync: async () => new Promise((res) => res('')),
-}))
-jest.mock('expo-gl', () => ({
-  GLView: ({ onContextCreate }: { onContextCreate: (gl: any) => void }) => {
-    React.useLayoutEffect(() => {
-      const gl = createWebGLContext({ width: 1280, height: 800 } as HTMLCanvasElement)
-      onContextCreate(gl)
-    }, [])
-
-    return null
-  },
-}))
